+++ src/view/components/CanvasComponentBase.js
